Replace option-arbitrary-content <selectlist> tests
This patch deletes some <selectlist> tests which test the arbitrary
rendering of <option>s. This behavior is already tested by the following
tests:
select-child-button-and-datalist-invalidation.tentative.html
select-child-button-and-datalist.tentative.html
select-open-invalidation.tentative.html

The deleted tests also test that images are rendered and that the
arbitrary rendering still works after the <option> is removed and
re-appended, so I added a new reftest which tests for these.

Fixed: 338513292
Bug: 335458492
Change-Id: Id1d5e1ffcfe9d2adecaada7188198fb933eca5c1